config.py
"""Centralized configuration for the FIH Rules RAG app.

Loads environment variables (from the OS and optionally .env) and exposes
typed constants for use across the codebase. Secrets such as DB_PASS are
required at runtime and not given insecure defaults.
"""

import os
try:
    from dotenv import load_dotenv  # type: ignore
except ImportError:
    load_dotenv = None

# Load environment variables from a local .env if present
if load_dotenv:
    load_dotenv()

# Google Cloud & Infra
PROJECT_ID = os.getenv("GCP_PROJECT_ID", "langchain-poc-479114")
REGION = os.getenv("GCP_REGION", "europe-west1")
# Document AI (US Fallback)
DOCAI_PROCESSOR_ID = os.getenv("DOCAI_PROCESSOR_ID", "53a6fece2d33c37")
DOCAI_LOCATION = os.getenv("DOCAI_LOCATION", "us")
# Staging bucket for Document AI
GCS_BUCKET_NAME = os.getenv("GCS_BUCKET_NAME", "fih-rag-staging-us-langchain-poc-479114")

# Database (Cloud SQL Postgres)
INSTANCE_NAME = os.getenv("CLOUDSQL_INSTANCE", "fih-rag-db")
DATABASE_NAME = "hockey_db"
TABLE_NAME = "hockey_rules_vectors"
DB_USER = os.getenv("DB_USER", "postgres")
DB_PASS = os.getenv("DB_PASS")  # Required; do not set a default here

# Model Config
EMBEDDING_MODEL = "text-embedding-004"
LLM_MODEL = "gemini-2.0-flash-lite"
RETRIEVAL_K = 15

# Supported Variants (key = DB label, value = UI label)
VARIANTS = {
    "outdoor": "Outdoor Hockey (Default)",
    "indoor": "Indoor Hockey",
    "hockey5s": "Hockey 5s"
}
